+++ src/components/BuyVault.jsx
@@ -1,0 +1,17 @@
+import './BuyVault.css';
+
+// Security modules
+import * as bip39 from '@scure/bip39';
+import { wordlist } from '@scure/bip39/wordlists/english';
+
+function BuyVault() {
+
+    return (
+        <>
+
+        </>
+    )
+
+}
+
+export default BuyVault;
